Create file "format_to_new.py" and move method "format_old_records_to_new" from "format.py" to "format_to_new.py"
